Fix Solomatin utils and decrease Java level yet again.

+++ src/main/java/ru/vsu/cs/util/DrawUtils.java
@@ -53,7 +53,7 @@ public class DrawUtils {
      * @return Контрастный цвет
      */
     public static Color getContrastColor(Color color) {
-        double y = (299 * color.getRed() + 587 * color.getGreen() + 114 * color.getBlue()) / 1000;
+        double y = (double) (299 * color.getRed() + 587 * color.getGreen() + 114 * color.getBlue()) / 1000;
         return y >= 128 ? Color.BLACK : Color.WHITE;
     }
 }
